+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * nearest_neighbours() function added
 
+* indicator_order() function added
+
 # fingertipsR 0.1.5 (06/02/2018)
 
 * corrected fingertips_stats to give accurate stats